Добавлен файли и папки для тесеов, исп по rootReducer и импорты
