+++ cypress/fixtures/user.json
@@ -1,0 +1,3 @@
+{
+  "example": "Using fixtures to represent data"
+}
