+++ src/components/burger-ingredient/burger-ingredient.tsx
@@ -3,7 +3,7 @@ import { useLocation } from 'react-router-dom';
 import { BurgerIngredientUI } from '@ui';
 import { TBurgerIngredientProps } from './type';
 import { useDispatch } from '../../services/store';
-import { addIngredient } from '../../services/slices/burgerconstructorSlice';
+import { addIngredient } from '../../services/slices/constructor/burgerconstructorSlice';
 
 export const BurgerIngredient: FC<TBurgerIngredientProps> = memo(
   ({ ingredient, count }) => {
