+++ src/pages/register/register.tsx
@@ -1,6 +1,6 @@
 import { FC, SyntheticEvent, useState } from 'react';
 import { RegisterUI } from '@ui-pages';
-import { registerUser } from '../../services/slices/userSlice';
+import { registerUser } from '../../services/slices/user/userSlice';
 import { useDispatch } from '../../services/store';
 import { useNavigate } from 'react-router-dom';
