+++ src/services/slices/constructor/index.ts
@@ -1,0 +1,1 @@
+export { burgerConstructorSlice } from './burgerconstructorSlice';
