+++ src/services/slices/feed/feedSlice.ts
@@ -32,7 +32,7 @@ const initialState: TFeedState = {
   userOrdersError: null
 };
 
-const feedSlice = createSlice({
+export const feedSlice = createSlice({
   name: 'feed',
   initialState,
   reducers: {},
