+++ src/services/slices/feed/index.ts
@@ -1,0 +1,1 @@
+export { feedSlice } from './feedSlice';
