+++ src/services/slices/order/index.ts
@@ -1,0 +1,1 @@
+export { orderSlice } from './orderSlice';
