+++ src/services/store.ts
@@ -1,32 +1,15 @@
-import { combineReducers, configureStore } from '@reduxjs/toolkit';
-import ingredientsSlice from './slices/ingredientsSlice';
-import burgerConstructorSlice from './slices/burgerconstructorSlice';
-import orderSlice from './slices/orderSlice';
+import { configureStore } from '@reduxjs/toolkit';
 import {
   TypedUseSelectorHook,
   useDispatch as dispatchHook,
   useSelector as selectorHook
 } from 'react-redux';
-import feedSlice from './slices/feedSlice';
-import userSlice from './slices/userSlice';
-const rootReducer = combineReducers({
-  // данные всех полученных ингредиентов;
-  ingredients: ingredientsSlice,
-  // данные конструктора бургера;
-  burgerConstructor: burgerConstructorSlice,
-  // данные создания заказа;
-  order: orderSlice,
-  // данные ленты и истории заказов;
-  feed: feedSlice,
-  // данные профиля пользователя.
-  user: userSlice
-});
+import rootReducer, { RootState } from './rootReducer';
+
 const store = configureStore({
   reducer: rootReducer,
   devTools: process.env.NODE_ENV !== 'production'
 });
-
-export type RootState = ReturnType<typeof store.getState>;
 
 export type AppDispatch = typeof store.dispatch;
